Re-architect library for better sharing
Move "host" libs from sut into the main library

Command to run on test box is now "hl"
Command to run on dev box is now "tl"

Default path is $HOME/tl; this is included in hl and tl (no more in .bashrc)

Make new "$gateway_override" command which will check to see if TEST_REMOTE
is true, and if so, ssh to the gateway instead.  Add this for wait-for-online
and wait-for-port commands.

Misc clean-up:
* Removed bogus "main" from bootlib.sh
* Add "pop" function to core.sh
* Moved "wait-for" commands to their own file
* Make "archive" for stuff I'm not using (but might want to keep around)
* Don't include xelib for now (since I can't really test it)

Signed-off-by: George Dunlap <george.dunlap@eu.citrix.com>