guestlib: Fix tgt caching
We want to avoid doing the whole recursive lookup thing if we already
have the information we want; but we also want to make sure to go
through the whole thing if it's actually changed.

Store the value of the last lookup; if the value is the same and we
already have the information we want, just return.  Otherwise, go
through the whole process.

Signed-off-by: George Dunlap <george.dunlap@eu.citrix.com>